requirements.txt
streamlit
fastapi
uvicorn
requests
sentence-transformers
torch
numpy
pandasfastapi
uvicorn
python-dotenv
stripe
pydantic